separate out Util module
